Remove -inl patterns in systems/analysis and move implementations to .cc files
