+++ systems/analysis/implicit_euler_integrator.h
@@ -4,8 +4,7 @@
 #include <memory>
 #include <utility>
 
-#include <Eigen/LU>
-
+#include "drake/common/default_scalars.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/math/autodiff_gradient.h"
 #include "drake/systems/analysis/implicit_integrator.h"
@@ -175,5 +174,9 @@ class ImplicitEulerIntegrator final : public ImplicitIntegrator<T> {
   int64_t num_err_est_jacobian_function_evaluations_{0};
   int64_t num_err_est_nr_iterations_{0};
 };
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class ImplicitEulerIntegrator)
+
 }  // namespace systems
 }  // namespace drake
