+++ systems/analysis/initial_value_problem.h
@@ -5,6 +5,7 @@
 #include <utility>
 #include <vector>
 
+#include "drake/common/default_scalars.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/common/eigen_types.h"
 #include "drake/systems/analysis/dense_output.h"
@@ -261,6 +262,9 @@ class InitialValueProblem {
   // Numerical integrator used for IVP ODE solving.
   std::unique_ptr<IntegratorBase<T>> integrator_;
 };
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class InitialValueProblem)
 
 }  // namespace systems
 }  // namespace drake
