--- systems/analysis/scalar_initial_value_problem-inl.h
@@ -1,5 +1,0 @@
-#pragma once
-
-// Instantiation of ScalarInitialValueProblem<T> for T other than double
-// requires InitialValueProblem<T> definition.
-#include "drake/systems/analysis/initial_value_problem-inl.h"
