+++ systems/analysis/scalar_initial_value_problem.cc
@@ -1,5 +1,4 @@
 #include "drake/systems/analysis/scalar_initial_value_problem.h"
-#include "drake/systems/analysis/scalar_initial_value_problem-inl.h"
 
 #include "drake/common/default_scalars.h"
 
@@ -7,7 +6,7 @@ namespace drake {
 namespace systems {
 
 DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
-    class ScalarInitialValueProblem);
+    class ScalarInitialValueProblem)
 
 }  // namespace systems
 }  // namespace drake
