+++ systems/analysis/simulator.h
@@ -12,6 +12,7 @@
 #include <vector>
 
 #include "drake/common/autodiff.h"
+#include "drake/common/default_scalars.h"
 #include "drake/common/drake_assert.h"
 #include "drake/common/drake_copyable.h"
 #include "drake/common/extract_double.h"
@@ -1589,6 +1590,9 @@ Simulator<T>::IntegrateContinuousState(
 
   DRAKE_UNREACHABLE();
 }
+
+DRAKE_DECLARE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class Simulator)
 
 }  // namespace systems
 }  // namespace drake
