+++ systems/analysis/test_utilities/controlled_spring_mass_system.cc
@@ -1,6 +1,7 @@
 #include "drake/systems/analysis/test_utilities/controlled_spring_mass_system.h"
 
 #include "drake/common/autodiff.h"
+#include "drake/common/default_scalars.h"
 #include "drake/common/eigen_types.h"
 #include "drake/systems/framework/diagram_builder.h"
 #include "drake/systems/primitives/demultiplexer.h"
@@ -118,8 +119,8 @@ const {
   return *plant_;
 }
 
-template class PidControlledSpringMassSystem<double>;
-template class PidControlledSpringMassSystem<AutoDiffXd>;
+DRAKE_DEFINE_CLASS_TEMPLATE_INSTANTIATIONS_ON_DEFAULT_NONSYMBOLIC_SCALARS(
+    class PidControlledSpringMassSystem)
 
 }  // namespace systems
 }  // namespace drake
